+++ src/dash-table/dash/DataTable.js
@@ -38,7 +38,7 @@ export const defaultProps = {
     css: [],
     filter_query: '',
     filter_action: 'none',
-    sort_as_none: [],
+    sort_as_null: [],
     sort_action: 'none',
     sort_mode: 'single',
     sort_by: [],
@@ -254,20 +254,20 @@ export const propTypes = {
         }),
 
         /**
-         * There are two `sort_as_none` flags in the table.
+         * There are two `sort_as_null` flags in the table.
          * This is the  column-level editable flag and there is
-         * also the table-level `sort_as_none` flag.
+         * also the table-level `sort_as_null` flag.
          *
          * These flags determine how the content of the table is
          * sorted.
          *
-         * If the column-level `sort_as_none` flag is set it overrides
-         * the table-level `sort_as_none` flag for that column.
+         * If the column-level `sort_as_null` flag is set it overrides
+         * the table-level `sort_as_null` flag for that column.
          *
          * An array of string, number and boolean values that are treated as `None`
          * when sorting is applied to the column.
          */
-        sort_as_none: PropTypes.arrayOf(PropTypes.oneOfType([
+        sort_as_null: PropTypes.arrayOf(PropTypes.oneOfType([
             PropTypes.string,
             PropTypes.number,
             PropTypes.bool
@@ -465,7 +465,7 @@ export const propTypes = {
             data: PropTypes.oneOf([0])
         }),
         PropTypes.exact({
-            headers: PropTypes.oneOf([true]),
+            headers: PropTypes.oneOf([true]).isRequired,
             data: PropTypes.number
         })
     ]),
@@ -894,11 +894,11 @@ export const propTypes = {
 
     /**
      * An array of string, number and boolean values that are treated as `None`
-     * when sorting. This value will be used by columns without `sort_as_none`.
+     * when sorting. This value will be used by columns without `sort_as_null`.
      *
      * Defaults to `[]`.
      */
-    sort_as_none: PropTypes.arrayOf(PropTypes.oneOfType([
+    sort_as_null: PropTypes.arrayOf(PropTypes.oneOfType([
         PropTypes.string,
         PropTypes.number,
         PropTypes.bool